+++ packages/pattern-commands/src/listeners/PluginMessageParse.ts
@@ -43,7 +43,9 @@ export class MessageParseListener extends Listener<typeof Events.PreMessageParse
 				return true;
 			}
 
-			if (content.match(new RegExp(key, client.options.caseInsensitiveCommands ? 'i' : undefined))) {
+			if (
+				content.match(new RegExp(patternCommand.matchFullName ? `\b${key}\b` : key, client.options.caseInsensitiveCommands ? 'i' : undefined))
+			) {
 				usedAlias = content;
 				return true;
 			}