+++ packages/scheduled-tasks/src/types-sqs.ts
@@ -1,0 +1,1 @@
+export type { ScheduledTaskSQSStrategy, ScheduledTaskSQSStrategyMessageBody, SQSClient } from './lib/strategies/ScheduledTaskSQSStrategy';