+++ packages/i18next/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.6](https://github.com/sapphiredev/plugins/compare/@sapphire/plugin-i18next@2.1.5...@sapphire/plugin-i18next@2.1.6) (2022-01-09)
+
+**Note:** Version bump only for package @sapphire/plugin-i18next
+
 ## [2.1.5](https://github.com/sapphiredev/plugins/compare/@sapphire/plugin-i18next@2.1.4...@sapphire/plugin-i18next@2.1.5) (2021-12-29)
 
 **Note:** Version bump only for package @sapphire/plugin-i18next